# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * ZeroRisco - Urban Mobility Platform API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export type UserInputRole = typeof UserInputRole[keyof typeof UserInputRole];


export const UserInputRole = {
  passenger: 'passenger',
  driver: 'driver',
} as const;

export interface UserInput {
  name: string;
  email: string;
  phone: string;
  password: string;
  role: UserInputRole;
  /** @nullable */
  vehicleModel?: string | null;
  /** @nullable */
  vehiclePlate?: string | null;
}

export interface LoginInput {
  email: string;
  password: string;
}

export interface UserProfile {
  id: number;
  name: string;
  email: string;
  phone: string;
  role: string;
  /** @nullable */
  avatarUrl?: string | null;
  /** @nullable */
  rating?: number | null;
  totalRides: number;
  createdAt: string;
}

export interface AuthResponse {
  token: string;
  user: UserProfile;
}

export interface CategoryInfo {
  id: number;
  name: string;
  description: string;
  icon: string;
  baseFare: number;
  pricePerKm: number;
  pricePerMinute: number;
  minFare: number;
  multiplier: number;
}

export interface RideEstimateInput {
  originLat: number;
  originLng: number;
  destinationLat: number;
  destinationLng: number;
}

export interface RideEstimate {
  categoryId: number;
  categoryName: string;
  categoryIcon: string;
  estimatedDistance: number;
  estimatedDuration: number;
  estimatedFare: number;
}

export interface RideInput {
  categoryId: number;
  originAddress: string;
  originLat: number;
  originLng: number;
  destinationAddress: string;
  destinationLat: number;
  destinationLng: number;
  estimatedDistance: number;
  estimatedDuration: number;
  estimatedFare: number;
}

export interface RideInfo {
  id: number;
  passengerId: number;
  /** @nullable */
  driverId?: number | null;
  categoryId: number;
  categoryName: string;
  categoryIcon: string;
  status: string;
  originAddress: string;
  originLat: number;
  originLng: number;
  destinationAddress: string;
  destinationLat: number;
  destinationLng: number;
  estimatedDistance: number;
  estimatedDuration: number;
  estimatedFare: number;
  /** @nullable */
  finalFare?: number | null;
  /** @nullable */
  driverName?: string | null;
  /** @nullable */
  driverPhone?: string | null;
  /** @nullable */
  driverRating?: number | null;
  /** @nullable */
  driverVehicle?: string | null;
  /** @nullable */
  driverPlate?: string | null;
  createdAt: string;
  /** @nullable */
  completedAt?: string | null;
}

export type RideStatusInputStatus = typeof RideStatusInputStatus[keyof typeof RideStatusInputStatus];


export const RideStatusInputStatus = {
  accepted: 'accepted',
  arrived: 'arrived',
  in_progress: 'in_progress',
  completed: 'completed',
  cancelled: 'cancelled',
} as const;

export interface RideStatusInput {
  status: RideStatusInputStatus;
  /** @nullable */
  driverId?: number | null;
}

export interface DriverInfo {
  id: number;
  userId: number;
  name: string;
  phone: string;
  vehicleModel: string;
  vehiclePlate: string;
  isOnline: boolean;
  rating: number;
  totalRides: number;
  balance: number;
}

export interface DriverStatusInput {
  isOnline: boolean;
}

export interface TransactionInfo {
  id: number;
  type: string;
  amount: number;
  description: string;
  createdAt: string;
}

export interface DriverEarnings {
  today: number;
  week: number;
  month: number;
  totalRides: number;
  balance: number;
  recentTransactions: TransactionInfo[];
}

export interface RatingInput {
  rideId: number;
  toUserId: number;
  rating: number;
  /** @nullable */
  comment?: string | null;
}

export interface RatingInfo {
  id: number;
  rideId: number;
  fromUserId: number;
  toUserId: number;
  rating: number;
  /** @nullable */
  comment?: string | null;
  createdAt: string;
}

export interface WalletInfo {
  balance: number;
  totalSpent: number;
  totalRides: number;
}

export interface MessageInfo {
  id: number;
  rideId: number;
  senderId: number;
  senderName: string;
  content: string;
  createdAt: string;
}

export interface MessageInput {
  content: string;
}

export interface NotificationInfo {
  id: number;
  title: string;
  body: string;
  type: string;
  isRead: boolean;
  createdAt: string;
}

export type ListRidesParams = {
status?: string;
limit?: number;
offset?: number;
};

export type ListTransactionsParams = {
limit?: number;
offset?: number;
};