# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: ZeroRisco - Urban Mobility Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: registerUser
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/login:
    post:
      operationId: loginUser
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user profile
      security:
        - bearerAuth: []
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /drivers/status:
    get:
      operationId: getDriverStatus
      tags: [drivers]
      summary: Get driver online/offline status and profile
      security:
        - bearerAuth: []
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DriverInfo"
    patch:
      operationId: updateDriverStatus
      tags: [drivers]
      summary: Toggle driver online/offline
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DriverStatusInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DriverInfo"
  /drivers/earnings:
    get:
      operationId: getDriverEarnings
      tags: [drivers]
      summary: Get driver earnings summary
      security:
        - bearerAuth: []
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DriverEarnings"
  /ratings:
    post:
      operationId: createRating
      tags: [ratings]
      summary: Rate a completed ride
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RatingInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RatingInfo"
    get:
      operationId: listRatings
      tags: [ratings]
      summary: List ratings for current user
      security:
        - bearerAuth: []
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/RatingInfo"
...
  /messages/{rideId}:
    get:
      operationId: listMessages
      tags: [messages]
      summary: List messages for a ride
      security:
        - bearerAuth: []
      parameters:
        - name: rideId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MessageInfo"
    post:
      operationId: sendMessage
      tags: [messages]
      summary: Send a message in a ride
      security:
        - bearerAuth: []
      parameters:
        - name: rideId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MessageInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageInfo"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
    UserInput:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        password:
          type: string
        role:
          type: string
          enum: [passenger, driver]
        vehicleModel:
          type: ["string", "null"]
        vehiclePlate:
          type: ["string", "null"]
      required: [name, email, phone, password, role]
    LoginInput:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
      required: [email, password]
    AuthResponse:
      type: object
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/UserProfile"
      required: [token, user]
    UserProfile:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        role:
          type: string
        avatarUrl:
          type: ["string", "null"]
        rating:
          type: ["number", "null"]
        totalRides:
          type: integer
        createdAt:
          type: string
      required: [id, name, email, phone, role, totalRides, createdAt]
...
    DriverInfo:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        name:
          type: string
        phone:
          type: string
        vehicleModel:
          type: string
        vehiclePlate:
          type: string
        isOnline:
          type: boolean
        rating:
          type: number
        totalRides:
          type: integer
        balance:
          type: number
      required: [id, userId, name, phone, vehicleModel, vehiclePlate, isOnline, rating, totalRides, balance]
    DriverStatusInput:
      type: object
      properties:
        isOnline:
          type: boolean
      required: [isOnline]
    DriverEarnings:
      type: object
      properties:
        today:
          type: number
        week:
          type: number
        month:
          type: number
        totalRides:
          type: integer
        balance:
          type: number
        recentTransactions:
          type: array
          items:
            $ref: "#/components/schemas/TransactionInfo"
      required: [today, week, month, totalRides, balance, recentTransactions]
    RatingInput:
      type: object
      properties:
        rideId:
          type: integer
        toUserId:
          type: integer
        rating:
          type: integer
        comment:
          type: ["string", "null"]
      required: [rideId, toUserId, rating]
    RatingInfo:
      type: object
      properties:
        id:
          type: integer
        rideId:
          type: integer
        fromUserId:
          type: integer
        toUserId:
          type: integer
        rating:
          type: integer
        comment:
          type: ["string", "null"]
        createdAt:
          type: string
      required: [id, rideId, fromUserId, toUserId, rating, createdAt]
...
    TransactionInfo:
      type: object
      properties:
        id:
          type: integer
        type:
          type: string
        amount:
          type: number
        description:
          type: string
        createdAt:
          type: string
      required: [id, type, amount, description, createdAt]
    MessageInfo:
      type: object
      properties:
        id:
          type: integer
        rideId:
          type: integer
        senderId:
          type: integer
        senderName:
          type: string
        content:
          type: string
        createdAt:
          type: string
      required: [id, rideId, senderId, senderName, content, createdAt]
    MessageInput:
      type: object
      properties:
        content:
          type: string
      required: [content]
...