# lib/api-zod/src/generated/types/driverInfo.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * ZeroRisco - Urban Mobility Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface DriverInfo {
  id: number;
  userId: number;
  name: string;
  phone: string;
  vehicleModel: string;
  vehiclePlate: string;
  isOnline: boolean;
  rating: number;
  totalRides: number;
  balance: number;
}