# lib/api-zod/src/generated/types/driverStatusInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * ZeroRisco - Urban Mobility Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface DriverStatusInput {
  isOnline: boolean;
}